Rework variables
