+++ game/Karen/Engine/Scripting/VariableContext.cs
@@ -15,20 +15,20 @@ namespace Karen.Engine.Scripting
 #if DEBUG
         public
 #endif
-        List<kvar> variables = new();
+        List<Variable> variables = new();
         public string Guid { get; private set; }
         public VariableContext()
         {
             Guid = Extensions.RandomString();
         }
-        public kvar Get(string name)
+        public Variable Get(string name)
         {
             foreach (var q in variables)
                 if (q.name == name)
                     return q;
             throw new ObjectNotFoundException($"In {Guid} context variable with {name} dont exist.");
         }
-        public void Add(kvar variable)
+        public void Add(Variable variable)
         {
             variables.Add(variable);
         }
